Almost Finish all / Services

+++ src/main/java/com/oxi/software/repository/PurchaseRepository.java
@@ -3,5 +3,5 @@ package com.oxi.software.repository;
 import com.oxi.software.entities.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseRepository extends JpaRepository<Purchase, Integer> {
+public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
 }
