+++ src/main/java/com/oxi/software/repository/RolTypeRepository.java
@@ -3,5 +3,5 @@ package com.oxi.software.repository;
 import com.oxi.software.entities.RolType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolTypeRepository extends JpaRepository<RolType, Integer> {
+public interface RolTypeRepository extends JpaRepository<RolType, Long> {
 }
