+++ src/main/java/com/oxi/software/repository/UnitRepository.java
@@ -3,5 +3,5 @@ package com.oxi.software.repository;
 import com.oxi.software.entities.Unit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnitRepository extends JpaRepository<Unit, Integer> {
+public interface UnitRepository extends JpaRepository<Unit, Long> {
 }
